improved iterative deepening

+++ boardEvaluate.h
@@ -15,7 +15,7 @@ enum {
     row, col, oblique_sum, oblique_delta
 };
 
-int64_t pos_estimate(int i, int j, int8_t player_side);
+int64_t pos_estimate(int i, int j, int player_side);
 
 int winner_check();
 
@@ -43,6 +43,6 @@ void ban_detect();
 
 void increment_void(int64_t v[][PATTERN_TYPES + 5], __AC_node *p, int32_t index);
 
-int64_t type_shift(int8_t type, int8_t player_side);
+int64_t type_shift(int8_t type, int player_side);
 
 #endif //GOBANG_BOARDEVALUATE_H
